Fix pch usage in oBase. Refactor and add a whole bunch of support to surface.h including conversions enough for oImage-comparable support.
Integrate lodepng and expose in-memory encode/decode. It's a lot slower both loading and saving than FreeImage but does compress better. Oh well, it's there. Revisit if the code savings/complexity is worth the effort, or perhaps rely on this as a fall-back and support FreeImage as a DLL.

Add image: a convenience layer on top of surface::buffer. The buffer can support any GPU-sampleable format with mips, cubes, slices, etc, but in most applications a simple 2D 1 slice no mip image will do. For all those limit the interface to only "normal" formats.