task_group + scheduler -> oBase/concurrency.h, similar to what used to be in oConcurrency.
Reintroduce separate libs for each concurrency implementation.

Use this to drastically reduce the dependencies of the exes, especially simpler tools.

Change linking to be librarian-based, not .obj-based so self-registration with a static won't work again... but put the setting back on for oUnitTests until all tests are migrated out of oTest's base class.

While reducing lib dependencies, I found a few dependencies that could be broken with some code reorg, so:
move stringize of half into oHLSL
move use of msgbox in reporting.cpp to be in oGUI and set at the start of each app where appropriate.